Code now checks for fighter.py and loads it instead of abstractFighter.
Code-based fighters still need a fighter.xml to hold their stats,
however.
