--- fighters/hitboxie/hitboxie.py
@@ -1,275 +1,0 @@
-import engine.abstractFighter as abstractFighter
-import settingsManager
-import os
-
-class Hitboxie(abstractFighter.AbstractFighter):
-    def __init__(self,playerNum,sprites):
-        var = {
-                'weight': 100,
-                'gravity': .5,
-                'maxFallSpeed': 20,
-                'maxGroundSpeed': 6,
-                'runSpeed': 9,
-                'maxAirSpeed': 6,
-                'crawlSpeed': 3,
-                'dodgeSpeed': 10,
-                'friction': 0.2,
-                'staticGrip': 0.1,
-                'airControl': 0.6,
-                'jumps': 1,
-                'jumpHeight': 12.5,
-                'shortHopHeight': 8.5,
-                'airJumpHeight': 15,
-                'heavyLandLag': 4,
-                'fastfallMultiplier': 2,
-                'hitstunElasticity': .8
-                }
-        abstractFighter.AbstractFighter.__init__(self,
-                                 playerNum,
-                                 sprites, #Start Sprite
-                                 "HBoxie", #Name
-                                 var)
-        self.actions = settingsManager.importFromURI(__file__,'hitboxie_actions.py',suffix=str(playerNum))
-        self.sideSpecialUses = 1
-        
-        #try:
-        self.current_action = self.actions.NeutralAction()
-        #except:
-        #    raise ValueError(os.path.normpath(os.path.join(os.path.dirname(__file__).replace('main.exe',''), 'hitboxie_actions.py')))
-        
-########################################################
-#                  ACTION SETTERS                      #
-########################################################
-    
-    def doIdle(self):
-        self.changeAction(self.actions.NeutralAction())
-
-    def doCrouch(self):
-        self.changeAction(self.actions.Crouch())
-
-    def doCrouchGetup(self):
-        self.changeAction(self.actions.CrouchGetup())
-    
-    def doFall(self):
-        self.changeAction(self.actions.Fall())
-
-    def doHelpless(self):
-        self.changeAction(self.actions.Helpless())
-    
-    def doPlatformDrop(self):
-        self.changeAction(self.actions.PlatformDrop())
-          
-    def doLand(self):
-        self.changeAction(self.actions.Land())
-
-    def doHelplessLand(self):
-        self.changeAction(self.actions.HelplessLand())
-        
-    def doStop(self):
-        if self.grounded:
-            self.changeAction(self.actions.Stop())
-
-    def doRunStop(self):
-        if self.grounded:
-            self.changeAction(self.actions.RunStop())
-            
-    def doGroundMove(self,direction):
-        if (self.facing == 1 and direction == 180) or (self.facing == -1 and direction == 0):
-            self.flip()
-        self.changeAction(self.actions.Move())
-
-    def doDash(self,direction):
-        if (self.facing == 1 and direction == 180) or (self.facing == -1 and direction == 0):
-            self.flip()
-        self.changeAction(self.actions.Dash())
-        
-    def doRun(self,direction):
-        if (self.facing == 1 and direction == 180) or (self.facing == -1 and direction == 0):
-            self.flip()
-        self.changeAction(self.actions.Run())
-        
-    def doPivot(self):
-        newAction = self.actions.Pivot()
-        self.flip()
-        self.changeAction(newAction)
-
-    def doRunPivot(self):
-        self.flip()
-        self.changeAction(self.actions.RunPivot())
-    
-    def doJump(self):
-        self.changeAction(self.actions.Jump())
-
-    def doAirJump(self):
-        self.changeAction(self.actions.AirJump())
-
-    def doPreShield(self):
-        self.changeAction(self.actions.PreShield())
-                
-    def doShield(self):
-        self.changeAction(self.actions.Shield())
-
-    def doShieldStun(self, length):
-        self.changeAction(self.actions.ShieldStun(length))
-        
-    def doForwardRoll(self):
-        self.changeAction(self.actions.ForwardRoll())
-    
-    def doBackwardRoll(self):
-        self.changeAction(self.actions.BackwardRoll())
-        
-    def doSpotDodge(self):
-        self.changeAction(self.actions.SpotDodge())
-        
-    def doAirDodge(self):
-        self.changeAction(self.actions.AirDodge())
-
-    def doTechDodge(self):
-        self.changeAction(self.actions.TechDodge())
-     
-    def doLedgeGrab(self,ledge):
-        self.changeAction(self.actions.LedgeGrab(ledge))
-
-    def doLedgeGetup(self):
-        self.changeAction(self.actions.LedgeGetup())
-
-    def doLedgeAttack(self):
-        self.changeAction(self.actions.LedgeAttack())
-
-    def doLedgeRoll(self):
-        self.changeAction(self.actions.LedgeRoll())
-
-    def doGroundGrab(self):
-        self.changeAction(self.actions.GroundGrab())
-
-    def doGrabbing(self):
-        self.changeAction(self.actions.Grabbing())
-
-    def doTrapped(self, length):
-        self.changeAction(self.actions.Trapped(length))
-
-    def doStunned(self, length):
-        self.changeAction(self.actions.Stunned(length))
-
-    def doGrabbed(self, height):
-        self.changeAction(self.actions.Grabbed(height))
-
-    def doPummel(self):
-        self.changeAction(self.actions.Pummel())
-
-    def doThrow(self):
-        (key, invkey) = self.getForwardBackwardKeys()
-        if self.keysContain(key):
-            self.changeAction(self.actions.ForwardThrow())
-        elif self.keysContain(invkey):
-            self.changeAction(self.actions.BackThrow())
-        elif self.keysContain('down'):
-            self.changeAction(self.actions.DownThrow())
-        elif self.keysContain('up'):
-            self.changeAction(self.actions.UpThrow())
-        else: # How did we get here? 
-            self.changeAction(self.actions.ForwardThrow())
-        
-    def doGroundAttack(self):
-        print('player ', self.playerNum, ' attacking')
-        (key, invkey) = self.getForwardBackwardKeys()
-        if self.keysContain(key):
-            if self.checkSmash(key):
-                print("SMASH!")
-                self.changeAction(self.actions.ForwardSmash())
-            else:
-                self.changeAction(self.actions.ForwardAttack())
-        elif self.keysContain(invkey):
-            self.flip()
-            if self.checkSmash(key):
-                print("SMASH!")
-                self.changeAction(self.actions.ForwardSmash())
-            else:
-                self.changeAction(self.actions.ForwardAttack())
-        elif self.keysContain('down'):
-            if self.checkSmash('down'):
-                print("SMASH!")
-                self.changeAction(self.actions.DownSmash())
-            else:
-                self.changeAction(self.actions.DownAttack())
-        elif self.keysContain('up'):
-            if self.checkSmash('up'):
-                print("SMASH!")
-                self.changeAction(self.actions.UpSmash())
-            else:
-                self.changeAction(self.actions.UpAttack())
-        else:
-            self.changeAction(self.actions.NeutralAttack())
-
-    def doDashAttack(self):
-        self.changeAction(self.actions.DashAttack())
-
-    def doDashGrab(self):
-        self.changeAction(self.actions.DashGrab())
-    
-    def doAirAttack(self):
-        (forward, backward) = self.getForwardBackwardKeys()
-        if (self.keysContain(forward)):
-            self.changeAction(self.actions.ForwardAir())
-        elif (self.keysContain(backward)):
-            self.changeAction(self.actions.BackAir())
-        elif (self.keysContain('down')):
-            self.changeAction(self.actions.DownAir())
-        elif(self.keysContain('up')):
-            self.changeAction(self.actions.UpAir())
-        else: self.changeAction(self.actions.NeutralAir())
-
-    def doGetupAttack(self, direction):
-        self.changeAction(self.actions.GetupAttack())
-    
-    def doGroundSpecial(self):
-        (forward, backward) = self.getForwardBackwardKeys()
-        if self.keysContain(forward):
-            if self.sideSpecialUses > 0:
-                self.changeAction(self.actions.ForwardSpecial())
-        elif self.keysContain(backward):
-            self.flip()
-            if self.sideSpecialUses > 0:
-                self.changeAction(self.actions.ForwardSpecial())
-        elif (self.keysContain('down')):
-            self.changeAction(self.actions.DownSpecial())
-        elif (self.keysContain('up')):
-            self.changeAction(self.actions.UpSpecial())
-        else: 
-            self.changeAction(self.actions.NeutralGroundSpecial())
-
-    def doAirSpecial(self):
-        (forward, backward) = self.getForwardBackwardKeys()
-        if self.keysContain(forward):
-            if self.sideSpecialUses > 0:
-                self.changeAction(self.actions.ForwardSpecial())
-        elif self.keysContain(backward):
-            self.flip()
-            if self.sideSpecialUses > 0:
-                self.changeAction(self.actions.ForwardSpecial())
-        elif (self.keysContain('down')):
-            self.changeAction(self.actions.DownSpecial())
-        elif (self.keysContain('up')):
-            self.changeAction(self.actions.UpSpecial())
-        else: 
-            self.changeAction(self.actions.NeutralAirSpecial())
-
-    def doHitStun(self,hitstun,trajectory,hitstop):
-        self.changeAction(self.actions.HitStun(hitstun,trajectory,hitstop))
-
-    def doTryTech(self, hitstun, trajectory, hitstop):
-        self.changeAction(self.actions.TryTech(hitstun, trajectory, hitstop))
-
-    def doTrip(self, length, direction):
-        self.changeAction(self.actions.Trip(length, direction))
-
-    def doGetup(self, direction):
-        self.changeAction(self.actions.Getup(direction))
-                
-########################################################
-#                  STATE CHANGERS                      #
-########################################################
-        
-    def die(self,respawn = True):
-        abstractFighter.AbstractFighter.die(self,respawn)
-        self.changeAction(self.actions.Fall())
