+++ fighters/sandbag/fighter.py
@@ -4,7 +4,7 @@ import os
 import settingsManager
 
 class Fighter(abstractFighter.AbstractFighter):
-    def __init__(self,playerNum):
+    def __init__(self,baseDir,playerNum):
         var = {
                 'weight': 200,
                 'gravity': 1.0,
@@ -21,39 +21,156 @@ class Fighter(abstractFighter.AbstractFighter):
                 'heavyLandLag': 4
                 }
         path = os.path.join(os.path.dirname(__file__),"sprites")
-        sprite = spriteManager.SpriteHandler(path,"","sandbag",128,{})
+        sprite = spriteManager.SpriteHandler(path,"sandbag_","idle",128,{})
         
-        abstractFighter.AbstractFighter.__init__(self,
-                                 playerNum,
-                                 sprite, #Start Sprite
-                                 "Sandbag", #Name
-                                 var) #jumps, jump height, air jump height
+        abstractFighter.AbstractFighter.__init__(self,baseDir,playerNum)
         
         self.actions = settingsManager.importFromURI(__file__,'sandbag_actions.py')
         
-        self.keyBindings = settingsManager.Keybindings({})
         self.current_action = self.actions.NeutralAction()
-        
-    def update(self):
-        abstractFighter.AbstractFighter.update(self)
-
-    #These are so grabbing the bag doesn't crash the game
-    def doGroundMove(self):
-        self.changeAction(self.actions.NeutralAction())
-
-    def doStop(self):
-        self.changeAction(self.actions.NeutralAction())
-
+    
+    
     def doIdle(self):
         self.changeAction(self.actions.NeutralAction())
 
-    def doGetup(self, direction):
-        self.changeAction(self.actions.NeutralAction())
+    def doCrouch(self):
+        pass
+
+    def doCrouchGetup(self):
+        pass
+    
+    def doFall(self):
+        self.changeAction(self.actions.Fall())
+
+    def doHelpless(self):
+        pass
+    
+    def doPlatformDrop(self):
+        pass
+          
+    def doLand(self):
+        pass
+
+    def doHelplessLand(self):
+        pass
+        
+    def doStop(self):
+        pass
+
+    def doRunStop(self):
+        pass
+            
+    def doGroundMove(self,direction):
+        pass
+
+    def doDash(self,direction):
+        pass
+        
+    def doRun(self,direction):
+        pass
+        
+    def doPivot(self):
+        pass
+
+    def doRunPivot(self):
+        pass
+    
+    def doJump(self):
+        pass
+
+    def doAirJump(self):
+        pass
+
+    def doPreShield(self):
+        pass
+                
+    def doShield(self):
+        pass
+
+    def doShieldStun(self, length):
+        pass
+        
+    def doForwardRoll(self):
+        pass
+    
+    def doBackwardRoll(self):
+        pass
+        
+    def doSpotDodge(self):
+        pass
+        
+    def doAirDodge(self):
+        pass
+
+    def doTechDodge(self):
+        pass
+     
+    def doLedgeGrab(self,ledge):
+        pass
+
+    def doLedgeGetup(self):
+        pass
+
+    def doLedgeAttack(self):
+        pass
+
+    def doLedgeRoll(self):
+        pass
+
+    def doGroundGrab(self):
+        pass
+
+    def doGrabbing(self):
+        pass
+
+    def doTrapped(self, length):
+        pass
+
+    def doStunned(self, length):
+        pass
+
+    def doGrabbed(self, height):
+        pass
+
+    def doPummel(self):
+        pass
+
+    def doThrow(self):
+        pass
+        
+    def doGroundAttack(self):
+        pass
+
+    def doDashAttack(self):
+        pass
+
+    def doDashGrab(self):
+        pass
+    
+    def doAirAttack(self):
+        pass
+
+    def doGetupAttack(self, direction):
+        pass
+    
+    def doGroundSpecial(self):
+        pass
+
+    def doAirSpecial(self):
+        pass
+
+    def doHitStun(self,hitstun,trajectory,hitstop):
+        pass
+
+    def doTryTech(self, hitstun, trajectory, hitstop):
+        pass
 
     def doTrip(self, length, direction):
-        self.changeAction(self.actions.Trip(length, direction))
-        
+        pass
+
+    def doGetup(self, direction):
+        pass    
 def cssIcon(): return None
 
-def getFighter(playerNum,colorNum):
-    return Fighter(playerNum)
+def getFighter(baseDir,playerNum):
+    return Fighter(baseDir,playerNum)
